backend/requirements.txt
fastapi
uvicorn
langchain
langchain-openai
langgraph
httpx
python-dotenv
pydantic
nest-asyncio
pandas
# mcp logic is implemented via httpx in tools.py
